# ResiduosApi/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ResiduosApi",
    "version": "v1"
  },
  "paths": {
    "/api/Residuos/ObtenerResiduos": {
      "get": {
        "tags": [
          "Residuos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Residuos/ObtenerResiduosXCiudad/{ciudad}": {
      "get": {
        "tags": [
          "Residuos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ciudad",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Residuos/ObtenerResiduosXCiudad": {
      "get": {
        "tags": [
          "Residuos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ciudad",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "valor",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Residuos/EnviarEvento": {
      "post": {
        "tags": [
          "Residuos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EventSend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Residuos/EventoHook": {
      "post": {
        "tags": [
          "Residuos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EventoWebHook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Payload": {
      "type": "object",
      "properties": {
        "titulo": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "valor": {
          "format": "int32",
          "type": "integer"
        },
        "flag": {
          "type": "boolean"
        }
      }
    },
    "EventSend": {
      "type": "object",
      "properties": {
        "event": {
          "type": "string"
        },
        "payload": {
          "$ref": "#/definitions/Payload"
        }
      }
    },
    "DataHook": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "sellerId": {
          "type": "string"
        },
        "sellerName": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "EventoWebHook": {
      "type": "object",
      "properties": {
        "eventType": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/DataHook"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  }
}